+++ client/src/components/volunteerManagement/VolunteerTopNav.tsx
@@ -1,43 +1,43 @@
-import { Box, Flex, Heading, Tabs } from "@chakra-ui/react";
-import { LuArchive, LuBriefcase, LuCircleUser } from "react-icons/lu";
-import { type ReactNode } from "react";
+import { useRef, useState } from "react";
 
+import { Box, Heading } from "@chakra-ui/react";
+
+import { FilterDrawer } from "./FilterDrawer";
 import { VolunteerManagementView } from "./VolunteerManagementView";
-
-const TABS: { value: string; icon: ReactNode; label: string }[] = [
-  { value: "archived", icon: <LuArchive />, label: "Archived" },
-  { value: "volunteers", icon: <LuCircleUser />, label: "Volunteer" },
-  { value: "staff", icon: <LuBriefcase />, label: "Staff" },
-];
+import { VolunteerToolbar } from "./VolunteerToolbar";
 
 export const VolunteerTopNav = () => {
+  const [filterDrawerOpen, setFilterDrawerOpen] = useState(false);
+  const [searchQuery, setSearchQuery] = useState("");
+  const [debouncedQuery, setDebouncedQuery] = useState("");
+  const debounceRef = useRef<ReturnType<typeof setTimeout> | null>(null);
+
+  const handleSearchChange = (e: React.ChangeEvent<HTMLInputElement>) => {
+    const val = e.target.value;
+    setSearchQuery(val);
+    if (debounceRef.current) clearTimeout(debounceRef.current);
+    debounceRef.current = setTimeout(() => setDebouncedQuery(val), 250);
+  };
+
   return (
     <Box p={6}>
-      <Heading size="2xl" m={4}>
+      <Heading size="2xl" mb={4}>
         User Management
       </Heading>
 
-      <Tabs.Root defaultValue="volunteers" variant="outline">
-        <Tabs.List>
-          {TABS.map(({ value, icon, label }) => (
-            <Tabs.Trigger key={value} value={value} color="#52525B" _selected={{ color: "#27272A", bg: "white" }}>
-              <Flex align="center" gap={1}>{icon} {label}</Flex>
-            </Tabs.Trigger>
-          ))}
-        </Tabs.List>
+      <VolunteerToolbar
+        searchQuery={searchQuery}
+        onSearchChange={handleSearchChange}
+        onFilterOpen={() => setFilterDrawerOpen(true)}
+      />
 
-        {/* TODO: Implement archived volunteers view */}
-        <Tabs.Content value="archived">
-          <p>archived!</p>
-        </Tabs.Content>
-        <Tabs.Content value="volunteers">
-          <VolunteerManagementView />
-        </Tabs.Content>
-        {/* TODO: Implement staff view */}
-        <Tabs.Content value="staff">
-          <p>staff!</p>
-        </Tabs.Content>
-      </Tabs.Root>
+      <VolunteerManagementView debouncedQuery={debouncedQuery} />
+
+      <FilterDrawer
+        open={filterDrawerOpen}
+        onClose={() => setFilterDrawerOpen(false)}
+        totalCount={0}
+      />
     </Box>
   );
 };
